Add TranscribeTask as primary runtime
* places the top level runtime that is currently distributed across
  whisperax_cli and the functions exported in whisperax.hpp in to one
  TranscribeTask()
* We can remove `extern` declaration from whisperax.hpp after we delete
  whisperax_cli target, as functions therein are used only by
  TranscribeTask

Moved AudioCodec related code to a private namespace and limited to
implementation file, since it is
required for a transcription task.
Moving to .cpp file before it is separated from TranscriptionTask means
we're not transitively including in places we don't need it.

Future work will likewise move this
to a separate file and abstraction, which will interface with both
real-time audio data and file inputs.